Suggest mode: source move-ghosts from the tree via a standalone hook
The pending-move ghost index was already computed from the live block tree
(metadata.suggestion.type === 'pending-move'), but that computation lived
inside SuggestionOverlayProvider and reached the rendering HOC through the
overlay context. Extract it verbatim into a standalone useMoveGhosts() hook
that depends only on the block-editor store, and have the block-class HOC call
it directly. Move suggestions live on the real block, so their ghosts must keep
rendering once the overlay is retired (#73411); this severs that dependency
without changing the data model or placement logic. All 327 suggestion-mode
unit tests still pass.
